"Applied fix from trunk for revision: 1846097" ------------------------------------------------------------------------ r1846097 | jleroux | 2018-11-08 08:18:43 +0100 (jeu. 08 nov. 2018) | 22 lignes
Fixed: Stores can't be modified at catalog/control/updateProductStore
(OFBIZ-10567)

This is related with the common-theme introduction.

There are 2 fields named visualThemeId in the data model. The one in 
ProductStore entity existed before the common-theme introduction with the new 
entity VisualTheme. When the theme is initialised with InitTheme.groovy it sets 
the context.visualThemeId value if it does not exist yet with 
parameters.visualThemeId. At this point things get confused and the 
backend theme used is replaced by the ecommerce theme defined for the 
productStore.

I think it would overwhelming to change the data model when only the 
EditProductStore form and updateProductStore service are concerned. 

So here is a very simple workaround where visualThemeId is replaced by 
ecomThemeId in the EditProductStore form and updateProductStore service.

If someone prefers to change the data model and related code feel free to revert, 
reopen the Jira and do.

------------------------------------------------------------------------


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/ofbiz-framework/branches/release17.12@1846098 13f79535-47bb-0310-9956-ffa450edef68