Documented: Backport the Asciidoc documentation and means to generate it to R17 branch (OFBIZ-10657)
As discussed in https://markmail.org/message/x7sv2mpvtytr4vl4 we decided to 
backport the Asciidoc documentation, and means to generate it, to R17 branch

This is related with OFBIZ-10651 and OFBIZ-10652 tasks. 
So 2 parts are impacted: code in framework and plusing and Buildbot script. 
It's only a copy from the trunk to the R17 branch, including *.adoc files, 
related images and the build.gradle script.

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/ofbiz-framework/branches/release17.12@1847248 13f79535-47bb-0310-9956-ffa450edef68