Applied fix from trunk for revision: 1848449 ===
Seed data sequence bug regarding "MIDNIGHT_DAILY" TemporalExpression.
(OFBIZ-10533)

The TemporalExpression for "MIDNIGHT_DAILY" has to be loaded as seed 
data instead of initial-seed data to be available for other seed-initial
usages of this TemporalExpression.
The fix moves this TemporalExpression to the other TemporalExpression 
seed data in ServiceSeedData.xml.

Thanks Martin Becker for reporting and providing the patch.


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/ofbiz-framework/branches/release17.12@1848450 13f79535-47bb-0310-9956-ffa450edef68