"Applied fix from trunk for revision: 1851319" ------------------------------------------------------------------------ r1851319 | jleroux | 2019-01-15 08:43:36 +0100 (mar. 15 janv. 2019) | 12 lignes
Fixed: partyId misses in EditEftAccount.ftl (party) when you create an EFT 
account from the quick finalize page
(OFBIZ-10680)

If you split the payment and want to create an EFT account you get an error in 
EditEftAccount.ftl (party) at line 40 colum 50: the partyId is missing.

Adds the patch to add the partyId field that was missing while creating an EFT 
account or credit card in split payment mode. Added proper checks to avoid the 
broken screen in case the parameter is missing.

Thanks: Ankit Joshi for the fix
------------------------------------------------------------------------


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/ofbiz-framework/branches/release18.12@1851320 13f79535-47bb-0310-9956-ffa450edef68