"Applied fix from trunk for revision: 1858180" ------------------------------------------------------------------------ r1858180 | jleroux | 2019-04-26 10:33:30 +0200 (ven. 26 avr. 2019) | 17 lignes
Reverted: Unknown request [images]; this request does not exist or cannot be 
called directly.
(OFBIZ-10895)

Partially revert r1858094, just replaces "rmaps == null" by "rmaps.isEmpty()"

As Mathieu mentioned in dev:
<<Checking for ‘null’ here is not a solution since ‘resolveURI’ contract
it to return a non-nullable collection.
This commit is only bypassing the error handling.>>

As I said in r1858094:
<<This has been somehow broken by OFBIZ-10438 but actually it was already broken 
before but not for the same reason.>>

We need now to review/check how it worked in R15 where it's OK, and possibly 
implements the same
------------------------------------------------------------------------


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/ofbiz-framework/branches/release18.12@1858181 13f79535-47bb-0310-9956-ffa450edef68