"Applied fix from trunk for revision: 1859576" ------------------------------------------------------------------------ r1859576 | jleroux | 2019-05-20 20:02:09 +0200 (lun. 20 mai 2019) | 15 lignes
Fixed: Unknown request [images]; this request does not exist or cannot be 
called directly.
(OFBIZ-10895)

This error happens in many occasions.
It shows only in log and have no effect on UI.
It happens in all supported releases. 
So it's not very bad since it's has no side effects.

As said Deepak Dixit:
I think we need to fix the image url in css file, it should start with /images, 
else browser append this in request info. There is nothing related to java or 
tomcat code. 

I finally decided to do a whole change, here it is
------------------------------------------------------------------------


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/ofbiz-framework/branches/release18.12@1859577 13f79535-47bb-0310-9956-ffa450edef68