"Applied fix from trunk for revision: 1863560 BY HAND" ------------------------------------------------------------------------ r1863560 | jleroux | 2019-07-22 17:21:14 +0200 (lun. 22 juil. 2019) | 37 lignes
Fixed: Unknown request [images]; this request does not exist or cannot be 
called directly.
(OFBIZ-10895)

Today Olivier reported directly to me that there was an issue in HR:
<<If you go to humanres, at the level of the tree that appears the icons to expand 
the tree no longer appear (but you can click...)
This is due to the path of the d.png image which is wrong in the css file 
(there is too much image in the path)>>

I should have seen it because the d.png file is, with other related image files,
in the same directory than style.css. And there was an inconsistency in the
style.css files regarding the d.png file path.

We should note that we have still this error in log:
2019-07-22 16:25:26,275 |jsse-nio-8443-exec-7 |ControlServlet                
|E| Error in request handler:
org.apache.ofbiz.webapp.control.RequestHandlerException: Unknown request [d.png]; 
this request does not exist or cannot be called directly.
at org.apache.ofbiz.webapp.control.RequestHandler.doRequest
(RequestHandler.java:277) ~[ofbiz.jar:?]

As I wrote in my last commit for OFBIZ-10895:
The only safe way to go is to not only check the error log for 
"Unknown request [...."
but also to look for 404 in access log, not all request show erroneously in
error log.

Some error appears before the user is logged in, as I wrote in the Jira:
<<I noticed that this error message often appears before login. A 404 can be 
seen in the access log and then once logged the issue no longer exists 
(ie we get a 200). I have no yet investigated why...>>

Nevertheless, I want to remove this misleading error message. I hope I'll do 
that before forgetting.

Thanks: Olivier Heintz 
------------------------------------------------------------------------


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/ofbiz-framework/branches/release17.12@1863563 13f79535-47bb-0310-9956-ffa450edef68